Week5

+++ src/solution/requests.cpp
@@ -1,8 +1,7 @@
 #include "requests.h"
+#include "transport_router.h"
 
 #include <vector>
-
-#include "transport_router.h"
 
 using namespace std;
 
@@ -33,7 +32,7 @@ namespace Requests {
       dict = {
           {"stop_count", Json::Node(static_cast<int>(bus->stop_count))},
           {"unique_stop_count", Json::Node(static_cast<int>(bus->unique_stop_count))},
-          {"route_length", Json::Node(bus->road_route_length)},
+          {"route_length", Json::Node(static_cast<int>(bus->road_route_length))},
           {"curvature", Json::Node(bus->road_route_length / bus->geo_route_length)},
       };
     }
@@ -42,10 +41,12 @@ namespace Requests {
 
   struct RouteItemResponseBuilder {
     Json::Dict operator()(const TransportRouter::RouteInfo::BusItem& bus_item) const {
-      return Json::Dict{{"type", Json::Node("Bus"s)},
-                        {"bus", Json::Node(bus_item.bus_name)},
-                        {"time", Json::Node(bus_item.time)},
-                        {"span_count", Json::Node(static_cast<int>(bus_item.span_count))}};
+      return Json::Dict{
+          {"type", Json::Node("Bus"s)},
+          {"bus", Json::Node(bus_item.bus_name)},
+          {"time", Json::Node(bus_item.time)},
+          {"span_count", Json::Node(static_cast<int>(bus_item.span_count))}
+      };
     }
     Json::Dict operator()(const TransportRouter::RouteInfo::WaitItem& wait_item) const {
       return Json::Dict{
@@ -100,12 +101,14 @@ namespace Requests {
     Json::Array responses;
     responses.reserve(requests.size());
     for (const Json::Node& request_node : requests) {
-      Json::Dict dict =
-          visit([&db](const auto& request) { return request.Process(db); }, Requests::Read(request_node.AsMap()));
+      Json::Dict dict = visit([&db](const auto& request) {
+                                return request.Process(db);
+                              },
+                              Requests::Read(request_node.AsMap()));
       dict["request_id"] = Json::Node(request_node.AsMap().at("id").AsInt());
       responses.push_back(Json::Node(dict));
     }
     return responses;
   }
 
-}  // namespace Requests
+}
